review bombing + hit events work
