+++ Content.Goobstation.Common/Events/UserShotAmmoEvent.cs
@@ -1,0 +1,17 @@
+using Robust.Shared.Serialization;
+
+namespace Content.Goobstation.Common.Events;
+
+/// <summary>
+/// Raised when entity shoots a gun.
+/// </summary>
+[Serializable, NetSerializable]
+public sealed class UserShotAmmoEvent : EntityEventArgs
+{
+    public List<NetEntity> FiredProjectiles { get; }
+
+    public UserShotAmmoEvent(List<NetEntity> firedProjectiles)
+    {
+        FiredProjectiles = firedProjectiles;
+    }
+}
