+++ Resources/Prototypes/_Goobstation/fonts.yml
@@ -1,0 +1,3 @@
+- type: font # goida
+  id: Style
+  path: /Fonts/_Goidastation/VCR_OSD_MONO_1.001.ttf
