+++ Resources/Prototypes/fonts.yml
@@ -53,7 +53,3 @@
 - type: font
   id: Emoji
   path: /Fonts/NotoEmoji.ttf
-
-- type: font # goidastation
-  id: Style
-  path: /Fonts/_Goidastation/VCR_OSD_MONO_1.001.ttf
